feat: Format results in the query panel
Now that we have externalized our small pp-display library we
can make the search panel depend on it and use it to display
results.

closes issue #748
