+++ client/pp-display/src/main.tsx
@@ -1,4 +1,6 @@
 import PpDisplay from './pp';
 import {PpString, PpMode, PpTag} from './types';
-export {PpDisplay};
+import {stringOfPpString} from './utilities';
+
+export {PpDisplay, stringOfPpString};
 export type {PpString, PpMode, PpTag};
