+++ client/search-ui/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@vscode/codicons": "^0.0.32",
     "@vscode/webview-ui-toolkit": "^1.2.1",
+    "pp-display": "link:../pp-display",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.7.1",
